Add support for proximity nearvalue and binarize quirk
Some proximity sensors don't report a distance.
Add support for forcing the value to be either near (0) or far (5).
This is done by enabling binarize quirk on the sensor.
The output value is "near" when the value is greater than nearlevel
which is read from sysfs and can be overriden by setting
ro.iio.prox.<name>.nearlevel property.

Signed-off-by: Vincent Knecht <vincent.knecht@mailoo.org>